+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
+- Lidarr metadata profile overrides service that creates, updates, and deletes metadata profiles to match `nixflix.lidarr.config.metadataProfiles` ([#293](https://github.com/kiriwalawren/nixflix/issues/293))
 - Lidarr quality profile overrides service that creates, updates, and deletes quality profiles to match `nixflix.lidarr.config.qualityProfiles` ([#294](https://github.com/kiriwalawren/nixflix/issues/294))
 - Navidrome ([#296](https://github.com/kiriwalawren/nixflix/pull/296))